优化了继承DEntityBase之后的代码 优化了枚举的使用 修改DbContext中获取用户id方法

+++ backend/Dilon.Core/Service/Auth/AuthService.cs
@@ -77,7 +77,7 @@ namespace Dilon.Core.Service
             _ = user ?? throw Oops.Oh(ErrorCode.D1000);
 
             // 验证账号是否被冻结
-            if (user.Status == (int)CommonStatus.DISABLE)
+            if (user.Status == CommonStatus.DISABLE)
                 throw Oops.Oh(ErrorCode.D1017);
 
             // 生成Token令牌
